fastapi_poe client: increase httpx timeout (#30)
If an httpx session is not passed in, the default timeout of the created session is currently 5 seconds, which is too short for long LLM requests.
Make the default 120 seconds instead. User can pass in their own httpx session if they wish to set a custom timeout.

Also, make the error log print repr(e) as httpx.ReadTimeout did not have a message body. So at least the exception type will be printed.