change attachment insertion method (#103)
Rather than concatenating file attachment content to last existing user message, insert them as new messages. This provides a better prompting for LLMs.
Will deprecate the old function, keeping it around for now. But this is still a breaking feature as the default has been changed to use the new insertion method instead.

* update the pyproject version

prepare 0.0.43 for release

* unit tests: remove 3.7 test from arm64

Was test fail with following message:
The version '3.7' with architecture 'arm64' was not found for macOS 14.4.1.
Which apparently is from github runner's macos-latest switching to an arm64 mac.