fastapi_poe: add bot_query_id (#114)
* fastapi_poe: add bot_query_id

this is a field that will be required for bot creators to pass along in
the point auth/capture requests as part of pay-by-context. this field
will be encrypted.

Reviewers:

Test Plan:

Differential Revision:

Asana Tasks:

Screenshot:

CC:

Deploy To:

* fastapi_poe: add bot_query_id

this is a field that will be required for bot creators to pass along in
the point auth/capture requests as part of pay-by-context. this field
will be encrypted.

Reviewers:

Test Plan:

Differential Revision:

Asana Tasks:

Screenshot:

CC:

Deploy To: