Bugfix/fix tool calling inefficiency (#147)
if tools are not selected in the first request with tools, openai will just stream back content.

previously behavior was to always make a second request after running tools locally.
This optimizes behavior to skip the second request and just return content from the first request if tools are not selected.